refactor: move edit reservation page
