+++ apps/admin-ui/src/spa/index.tsx
