+++ apps/admin-ui/src/spa/reservations/[id]/RecurringReservationsView.tsx
@@ -12,7 +12,7 @@ import ReservationListButton from "@component/ReservationListButton";
 import { useModal } from "@context/ModalContext";
 import DenyDialog from "./DenyDialog";
 import { useRecurringReservations } from "@hooks/reservations";
-import EditTimeModal from "../EditTimeModal";
+import EditTimeModal from "./edit/EditTimeModal";
 
 export function RecurringReservationsView({
   recurringPk,
