+++ apps/admin-ui/src/spa/reservations/[id]/edit/EditTimeModal.tsx
@@ -25,9 +25,10 @@ import { ControlledTimeInput } from "@component/ControlledTimeInput";
 import { ControlledDateInput } from "@component/ControlledDateInput";
 import BufferToggles from "@component/my-units/BufferToggles";
 import { useCheckCollisions } from "@hooks/reservations";
-import { CHANGE_RESERVATION_TIME } from "./queries";
 import { setTimeOnDate } from "@common/util";
-import { reservationDateTime, reservationDuration } from "./[id]/util";
+// TODO move
+import { CHANGE_RESERVATION_TIME } from "../../queries";
+import { reservationDateTime, reservationDuration } from "../util";
 
 const StyledForm = styled.form`
   margin-top: var(--spacing-m);
